gcum docstring
I worked on the gcum functions that calculate cumulative probability of the standard normal distribution. We are given parameter x, which is the value of standard norm distribution. This probability is the area under the normal curve to the left. The given 'x' should be between the values -3 - 6 for the return value to be 0.00 - 1.00. Anything greater than 6 is 1.00 and anything less than -3 is 0.00 for the return value. The return value is the cumlative probability which is returned as a float. The parameter 'x' can be a float or a int. There are two gcum functions. Even though the descriptions before I started the docstring are different, both functions calculate the same thing and do the same thing.